style: add some docstrings and improve structure
