+++ src/components/shared/copyright.js
@@ -1,0 +1,22 @@
+import Typography from "@mui/material/Typography";
+import Link from "@mui/material/Link";
+
+const Copyright = (props) => {
+  return (
+    <Typography
+      variant="body2"
+      color="text.secondary"
+      align="center"
+      {...props}
+    >
+      {"Copyright © "}
+      <Link color="inherit" href="#">
+        Blogs
+      </Link>{" "}
+      {new Date().getFullYear()}
+      {"."}
+    </Typography>
+  );
+};
+
+export default Copyright;
